Remove root wrapper module
