+++ features/F4/home_index_module/__init__.py
@@ -1,3 +1,3 @@
-from .run_server import run_server
+from .run_server import run_server, segments_to_chunk_docs, split_chunk_docs
 
-__all__ = ["run_server"]
+__all__ = ["run_server", "segments_to_chunk_docs", "split_chunk_docs"]
