+++ features/F5/__init__.py
@@ -1,5 +1,6 @@
 """Feature F5 package."""
 
 from . import chunk_module
+from . import chunk_utils
 
-__all__ = ["chunk_module"]
+__all__ = ["chunk_module", "chunk_utils"]
